pyproject.toml
[tool.poetry]
name = "python-paseto"
version = "0.1.0"
description = "Platform-Agnostic Security Tokens for Python"
authors = []
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6 || ^3.7"
pynacl = "^1.3"

[tool.poetry.dev-dependencies]
pytest = "^3.9"
pytest-cov = "^2.6"
black = "^18.3-alpha.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
